Remove two files. Revise c.beautify_script_tree. Add g.beautify_with_ruff
