+++ .mypy.ini
@@ -26,7 +26,7 @@ exclude =
     doc/|dist/|editpane/|examples/|extensions/|external/|modes/|obsolete/|scripts/|themes/|unittests/|www/|
     
     # Special plugins...
-    |freewin.py|scriptFile.py|viewrendered3|
+    |freewin.py|scriptFile.py|beautify_node.py|viewrendered3|
 
     # Experimental plugins...
     |cursesGui|leo_pdf|leo_to_rtf|mnplugins|mod_speedups|settings_finder|stickynotes_plus|
