+++ leo/unittests/core/test_leoAst.py
@@ -14,7 +14,6 @@ import time
 import token as token_module
 from typing import Any
 import unittest
-import warnings
 
 # Third-party.
 try:
@@ -22,22 +21,23 @@ try:
 except Exception:  # pragma: no cover
     asttokens = None
 
-try:
-    # Suppress a warning about imp being deprecated.
-    with warnings.catch_warnings():
-        import black
-except Exception:  # pragma: no cover
-    black = None
+# import warnings
+# try:
+#     # Suppress a warning about imp being deprecated.
+#     with warnings.catch_warnings():
+#     import black
+# except Exception:  # pragma: no cover
+#     black = None
 
 from leo.core import leoGlobals as g
 
 from leo.core.leoAst import AstNotEqual
-from leo.core.leoAst import Fstringify, Orange
+from leo.core.leoAst import Fstringify
 from leo.core.leoAst import Token, TokenOrderGenerator
 from leo.core.leoAst import make_tokens, parse_ast, tokens_to_string
 from leo.core.leoAst import dump_ast, dump_contents, dump_tokens, dump_tree, _op_names
 
-warnings.simplefilter("ignore")
+# warnings.simplefilter("ignore")
 # @-<< test_leoAst imports >>
 v1, v2, junk1, junk2, junk3 = sys.version_info
 py_version = (v1, v2)
@@ -345,30 +345,6 @@ class BaseTest(unittest.TestCase):
         result_s = fs.fstringify(contents, filename, tokens, tree)
         t2 = get_time()
         self.update_times('21: fstringify', t2 - t1)
-        return result_s
-
-    # @+node:ekr.20200107175223.1: *5* 2.2: BaseTest.beautify
-    def beautify(self, contents, tokens, tree, filename=None, max_join_line_length=None, max_split_line_length=None):
-        """
-        BaseTest.beautify.
-        """
-        t1 = get_time()
-        if not contents:
-            return ''  # pragma: no cover
-        if not filename:
-            filename = g.callers(2).split(',')[0]
-        orange = Orange()
-        result_s = orange.beautify(
-            contents,
-            filename,
-            tokens,
-            tree,
-            max_join_line_length=max_join_line_length,
-            max_split_line_length=max_split_line_length,
-        )
-        t2 = get_time()
-        self.update_times('22: beautify', t2 - t1)
-        self.code_list = orange.code_list
         return result_s
 
     # @+node:ekr.20191228095945.1: *4* BaseTest: stats...
@@ -1811,943 +1787,6 @@ class TestFstringify(BaseTest):
         expected = contents
         results = self.fstringify(contents, tokens, tree)
         self.assertEqual(results, expected)
-
-    # @-others
-
-
-# @+node:ekr.20200107174645.1: *3* class TestOrange (BaseTest)
-class TestOrange(BaseTest):
-    """
-    Tests for the Orange class.
-
-    **Important**: All unit tests assume that black_mode is False.
-                   That is, unit tests assume that no blank lines
-                   are ever inserted or deleted.
-    """
-
-    # @+others
-    # @+node:ekr.20200115201823.1: *4* TestOrange.blacken
-    def blacken(self, contents, line_length=None):
-        """Return the results of running black on contents"""
-        if not black:
-            self.skipTest('Requires Black')  # pragma: no cover
-        # Suppress string normalization!
-        try:
-            mode = black.FileMode()
-            mode.string_normalization = False
-            if line_length is not None:
-                mode.line_length = line_length
-        except TypeError:  # pragma: no cover
-            self.skipTest('Requires later version of Black')
-        return black.format_str(contents, mode=mode)
-
-    # @+node:ekr.20230115150916.1: *4* TestOrange.test_annotations
-    def test_annotations(self):
-        table = (
-            # Case 0.
-            '''
-    def annotated_f(s: str = None, x=None) -> None:
-        pass
-    ''',
-        )
-        for i, contents in enumerate(table):
-            contents, tokens, tree = self.make_data(contents)
-            expected = self.blacken(contents).strip() + '\n'
-            results = self.beautify(contents, tokens, tree)
-            self.assertEqual(results, expected)
-
-    # @+node:ekr.20200219114415.1: *4* TestOrange.test_at_doc_part
-    def test_at_doc_part(self):
-        line_length = 40  # For testing.
-        contents = """
-    # @+at Line 1
-    # Line 2
-    # @@c
-
-    print('hi')
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents.rstrip() + '\n'
-        results = self.beautify(
-            contents,
-            tokens,
-            tree,
-            max_join_line_length=line_length,
-            max_split_line_length=line_length,
-        )
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200116102345.1: *4* TestOrange.test_backslash_newline
-    def test_backslash_newline(self):
-        """
-        This test is necessarily different from black, because orange doesn't
-        delete semicolon tokens.
-        """
-        contents = r"""
-    print(a);\
-    print(b)
-    print(c); \
-    print(d)
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents.rstrip() + '\n'
-        # expected = self.blacken(contents).rstrip() + '\n'
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200219145639.1: *4* TestOrange.test_blank_lines_after_function
-    def test_blank_lines_after_function(self):
-        contents = """
-    # Comment line 1.
-    # Comment line 2.
-
-    def spam():
-        pass
-        # Properly indented comment.
-
-    # Comment line3.
-    # Comment line4.
-    a = 2
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200220050758.1: *4* TestOrange.test_blank_lines_after_function_2
-    def test_blank_lines_after_function_2(self):
-        contents = """
-    # Leading comment line 1.
-    # Leading comment lines 2.
-
-    def spam():
-        pass
-
-    # Trailing comment line.
-    a = 2
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200220053212.1: *4* TestOrange.test_blank_lines_after_function_3
-    def test_blank_lines_after_function_3(self):
-        # From leoAtFile.py.
-        contents = """
-    def writeAsisNode(self, p):
-        print('1')
-
-        def put(s):
-            print('2')
-
-        # Trailing comment 1.
-        # Trailing comment 2.
-        print('3')
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200228074455.1: *4* TestOrange.test_bug_1429
-    def test_bug_1429(self):
-        contents = r'''
-    def get_semver(tag):
-        """bug 1429 docstring"""
-        try:
-            import semantic_version
-            version = str(semantic_version.Version.coerce(tag, partial=True))
-                # tuple of major, minor, build, pre-release, patch
-                # 5.6b2 --> 5.6-b2
-        except(ImportError, ValueError) as err:
-            print('\n', err)
-            print("""*** Failed to parse Semantic Version from git tag '{0}'.
-            Expecting tag name like '5.7b2', 'leo-4.9.12', 'v4.3' for releases.
-            This version can't be uploaded to PyPi.org.""".format(tag))
-            version = tag
-        return version
-    '''
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents.strip() + '\n'
-        results = self.beautify(contents, tokens, tree, max_join_line_length=0, max_split_line_length=0)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20210318055702.1: *4* TestOrange.test_bug_1851
-    def test_bug_1851(self):
-        contents = r'''
-    def foo(a1):
-        pass
-    '''
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents.strip() + '\n'
-        results = self.beautify(contents, tokens, tree, max_join_line_length=0, max_split_line_length=0)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200209152745.1: *4* TestOrange.test_comment_indented
-    def test_comment_indented(self):
-        line_length = 40  # For testing.
-        table = (
-            """
-    if 1:
-        pass
-            # An indented comment.
-    """,
-            """
-    table = (
-        # Indented comment.
-    )
-    """,
-        )
-
-        fails = 0
-        for contents in table:
-            contents, tokens, tree = self.make_data(contents)
-            expected = contents
-            if 0:
-                dump_contents(contents)
-                dump_tokens(tokens)
-                # dump_tree(tokens, tree)
-            results = self.beautify(
-                contents,
-                tokens,
-                tree,
-                max_join_line_length=line_length,
-                max_split_line_length=line_length,
-            )
-            message = f"\n  contents: {contents!r}\n  expected: {expected!r}\n       got: {results!r}"
-            if results != expected:  # pragma: no cover
-                fails += 1
-                print(f"Fail: {fails}\n{message}")
-        assert not fails, fails
-
-    # @+node:ekr.20230117043931.1: *4* TestOrange.test_comment_space_after_delim
-    def test_comment_space_after_delim(self):
-        line_length = 40  # For testing.
-        table = (
-            # Test 1.
-            (
-                """#No space after delim.\n""",
-                """# No space after delim.\n""",
-            ),
-            # Test 2.  Don't change bang lines.
-            (
-                """#! /usr/bin/env python\n""",
-                """#! /usr/bin/env python\n""",
-            ),
-            # Test 3.  Don't change ### comments.
-            (
-                """### To do.\n""",
-                """### To do.\n""",
-            ),
-        )
-        fails = 0
-        for contents, expected in table:
-            contents, tokens, tree = self.make_data(contents)
-            if 0:
-                dump_contents(contents)
-                dump_tokens(tokens)
-                # dump_tree(tokens, tree)
-            results = self.beautify(
-                contents,
-                tokens,
-                tree,
-                max_join_line_length=line_length,
-                max_split_line_length=line_length,
-            )
-            message = f"\n  contents: {contents!r}\n  expected: {expected!r}\n       got: {results!r}"
-            if results != expected:  # pragma: no cover
-                fails += 1
-                print(f"Fail: {fails}\n{message}")
-        assert not fails, fails
-
-    # @+node:ekr.20200210120455.1: *4* TestOrange.test_decorator
-    def test_decorator(self):
-        table = (
-            # Case 0.
-            """
-    @my_decorator(1)
-    def func():
-        pass
-    """,
-            # Case 1.
-            """
-    if 1:
-        @my_decorator
-        def func():
-            pass
-    """,
-            # Case 2.
-            '''
-    @g.commander_command('promote')
-    def promote(self, event=None, undoFlag=True):
-        """Make all children of the selected nodes siblings of the selected node."""
-    ''',
-        )
-        for i, contents in enumerate(table):
-            contents, tokens, tree = self.make_data(contents)
-            expected = self.prep(contents)
-            results = self.beautify(contents, tokens, tree)
-            if results != expected:
-                g.trace('Fail:', i)  # pragma: no cover
-            self.assertEqual(results, expected)
-
-    # @+node:ekr.20200211094614.1: *4* TestOrange.test_dont_delete_blank_lines
-    def test_dont_delete_blank_lines(self):
-        line_length = 40  # For testing.
-        contents = """
-    class Test:
-
-        def test_func():
-
-            pass
-
-        a = 2
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents.rstrip() + '\n'
-        results = self.beautify(
-            contents,
-            tokens,
-            tree,
-            max_join_line_length=line_length,
-            max_split_line_length=line_length,
-        )
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200116110652.1: *4* TestOrange.test_function_defs
-    def test_function_defs(self):
-        table = (
-            # Case 0.
-            """
-    def f1(a=2 + 5):
-        pass
-    """,
-            # Case 2
-            """
-    def f1():
-        pass
-    """,
-            # Case 3.
-            """
-    def f1():
-        pass
-    """,
-            # Case 4.
-            '''
-    def should_kill_beautify(p):
-        """Return True if p.b contains @killbeautify"""
-        return 'killbeautify' in g.get_directives_dict(p)
-    ''',
-        )
-        for i, contents in enumerate(table):
-            contents, tokens, tree = self.make_data(contents)
-            expected = self.blacken(self.prep(contents))
-            results = self.beautify(contents, tokens, tree)
-            self.assertEqual(results, expected)
-
-    # @+node:ekr.20200116104031.1: *4* TestOrange.test_join_and_strip_condition
-    def test_join_and_strip_condition(self):
-        contents = self.prep(
-            """
-            if (
-                a == b or
-                c == d
-            ):
-                pass
-        """
-        )
-        expected = self.prep(
-            """
-            if (a == b or c == d):
-                pass
-        """
-        )
-        contents, tokens, tree = self.make_data(contents)
-        expected = self.prep(expected)
-        # Black also removes parens, which is beyond our scope at present.
-        # expected = self.blacken(contents, line_length=40)
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200208041446.1: *4* TestOrange.test_join_leading_whitespace
-    def test_join_leading_whitespace(self):
-        line_length = 40  # For testing.
-        table = (
-            # 1234567890x1234567890x1234567890x1234567890x
-            """
-    if 1:
-        print('4444',
-            '5555')
-    """,
-            """
-    if 1:
-        print('4444', '5555')\n""",
-        )
-        fails = 0
-        for contents in table:
-            contents, tokens, tree = self.make_data(contents)
-            if 0:
-                dump_contents(contents)
-                dump_tokens(tokens)
-                # dump_tree(tokens, tree)
-            expected = contents
-            # expected = self.blacken(contents, line_length=line_length)
-            results = self.beautify(
-                contents,
-                tokens,
-                tree,
-                max_join_line_length=line_length,
-                max_split_line_length=line_length,
-            )
-            message = f"\n  contents: {contents!r}\n  expected: {expected!r}\n       got: {results!r}"
-            if results != expected:  # pragma: no cover
-                fails += 1
-                print(f"Fail: {fails}\n{message}")
-        assert not fails, fails
-
-    # @+node:ekr.20200121093134.1: *4* TestOrange.test_join_lines
-    def test_join_lines(self):
-        # Except where noted, all entries are expected values....
-        line_length = 40  # For testing.
-        table = (
-            # 1234567890x1234567890x1234567890x1234567890x
-            """print('4444',\n    '5555')""",
-            """print('4444', '5555')\n""",
-        )
-        fails = 0
-        for contents in table:
-            contents, tokens, tree = self.make_data(contents)
-            if 0:
-                dump_contents(contents)
-                dump_tokens(tokens)
-                # dump_tree(tokens, tree)
-            expected = contents
-            results = self.beautify(
-                contents,
-                tokens,
-                tree,
-                max_join_line_length=line_length,
-                max_split_line_length=line_length,
-            )
-            message = f"\n  contents: {contents!r}\n  expected: {expected!r}\n    orange: {results!r}"
-            if results != expected:  # pragma: no cover
-                fails += 1
-                print(f"Fail: {fails}\n{message}")
-        self.assertEqual(fails, 0)
-
-    # @+node:ekr.20200210051900.1: *4* TestOrange.test_join_suppression
-    def test_join_suppression(self):
-        contents = self.prep("""
-            class T:
-                a = 1
-                print(
-                   a
-                )
-        """)
-        expected = self.prep("""
-            class T:
-                a = 1
-                print(a)
-        """)
-        contents, tokens, tree = self.make_data(contents)
-        expected = self.prep(expected)
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200207093606.1: *4* TestOrange.test_join_too_long_lines
-    def test_join_too_long_lines(self):
-        # Except where noted, all entries are expected values....
-        line_length = 40  # For testing.
-        table = (
-            # 1234567890x1234567890x1234567890x1234567890x
-            (
-                """print('aaaaaaaaaaaa',\n    'bbbbbbbbbbbb', 'cccccccccccccccc')""",
-                """print('aaaaaaaaaaaa',\n    'bbbbbbbbbbbb', 'cccccccccccccccc')\n""",
-            ),
-        )
-        fails = 0
-        for contents, expected in table:
-            contents, tokens, tree = self.make_data(contents)
-            if 0:
-                dump_contents(contents)
-                dump_tokens(tokens)
-                # dump_tree(tokens, tree)
-            results = self.beautify(
-                contents,
-                tokens,
-                tree,
-                max_join_line_length=line_length,
-                max_split_line_length=line_length,
-            )
-            message = f"\n  contents: {contents!r}\n  expected: {expected!r}\n       got: {results!r}"
-            if results != expected:  # pragma: no cover
-                fails += 1
-                print(f"Fail: {fails}\n{message}")
-        assert not fails, fails
-
-    # @+node:ekr.20220327131225.1: *4* TestOrange.test_leading_stars
-    def test_leading_stars(self):
-        # #2533.
-        contents = (
-            """
-            def f(
-                arg1,
-                *args,
-                **kwargs
-            ):
-                pass
-        """.strip()
-            + '\n'
-        )
-
-        expected = (
-            """
-            def f(arg1, *args, **kwargs):
-                pass
-        """.strip()
-            + '\n'
-        )
-        contents, tokens, tree = self.make_data(contents)
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(expected, results)
-
-    # @+node:ekr.20200108075541.1: *4* TestOrange.test_leo_sentinels
-    def test_leo_sentinels_1(self):
-        # Careful: don't put a sentinel into the file directly.
-        # That would corrupt leoAst.py.
-        sentinel = '#@+node:ekr.20200105143308.54: ** test'
-        contents = f"""
-    {sentinel}
-    def spam():
-        pass
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents.rstrip() + '\n'
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200209155457.1: *4* TestOrange.test_leo_sentinels_2
-    def test_leo_sentinels_2(self):
-        # Careful: don't put a sentinel into the file directly.
-        # That would corrupt leoAst.py.
-        sentinel = '#@+node:ekr.20200105143308.54: ** test'
-        contents = f"""
-    {sentinel}
-    class TestClass:
-        pass
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents.rstrip() + '\n'
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200108082833.1: *4* TestOrange.test_lines_before_class
-    def test_lines_before_class(self):
-        contents = """
-    a = 2
-    class aClass:
-        pass
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200110014220.86: *4* TestOrange.test_multi_line_pet_peeves
-    def test_multi_line_pet_peeves(self):
-        contents = """
-    if x == 4: pass
-    if x == 4 : pass
-    print (x, y); x, y = y, x
-    print (x , y) ; x , y = y , x
-    if(1):
-        pass
-    elif(2):
-        pass
-    while(3):
-        pass
-    """
-        # At present Orange doesn't split lines...
-        expected = """
-    if x == 4: pass
-    if x == 4: pass
-    print(x, y); x, y = y, x
-    print(x, y); x, y = y, x
-    if (1):
-        pass
-    elif (2):
-        pass
-    while (3):
-        pass
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = self.adjust_expected(expected)
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200110014220.95: *4* TestOrange.test_one_line_pet_peeves
-    def test_one_line_pet_peeves(self):
-        # See https://peps.python.org/pep-0008/#pet-peeves
-        # See https://peps.python.org/pep-0008/#other-recommendations
-
-        tag = 'test_one_line_pet_peeves'
-        # Except where noted, all entries are expected values....
-        if 0:
-            # Test fails or recents...
-            table = (
-                # """a[: 1 if True else 2 :]""",
-                """a[:-1]""",
-            )
-        else:
-            table = (
-                # Assignments...
-                # Slices (colons)...
-                """a[:-1]""",
-                """a[: 1 if True else 2 :]""",
-                """a[1 : 1 + 2]""",
-                """a[lower:]""",
-                """a[lower::]""",
-                """a[:upper]""",
-                """a[:upper:]""",
-                """a[::step]""",
-                """a[lower:upper:]""",
-                """a[lower:upper:step]""",
-                """a[lower + offset : upper + offset]""",
-                """a[: upper_fn(x) :]""",
-                """a[: upper_fn(x) : step_fn(x)]""",
-                """a[:: step_fn(x)]""",
-                """a[: upper_fn(x) :]""",
-                """a[: upper_fn(x) : 2 + 1]""",
-                """a[:]""",
-                """a[::]""",
-                """a[1:]""",
-                """a[1::]""",
-                """a[:2]""",
-                """a[:2:]""",
-                """a[::3]""",
-                """a[1:2]""",
-                """a[1:2:]""",
-                """a[:2:3]""",
-                """a[1:2:3]""",
-                # * and **, inside and outside function calls.
-                """a = b * c""",
-                # Now done in test_star_star_operator
-                # """a = b ** c""",  # Black has changed recently.
-                """f(*args)""",
-                """f(**kwargs)""",
-                """f(*args, **kwargs)""",
-                """f(a, *args)""",
-                """f(a=2, *args)""",
-                # Calls...
-                """f(-1)""",
-                """f(-1 < 2)""",
-                """f(1)""",
-                """f(2 * 3)""",
-                """f(2 + name)""",
-                """f(a)""",
-                """f(a.b)""",
-                """f(a=2 + 3, b=4 - 5, c= 6 * 7, d=8 / 9, e=10 // 11)""",
-                """f(a[1 + 2])""",
-                """f({key: 1})""",
-                """t = (0,)""",
-                """x, y = y, x""",
-                # Dicts...
-                """d = {key: 1}""",
-                """d['key'] = a[i]""",
-                # Trailing comments: expect two spaces.
-                """whatever # comment""",
-                """whatever  # comment""",
-                """whatever   # comment""",
-                # Word ops...
-                """v1 = v2 and v3 if v3 not in v4 or v5 in v6 else v7""",
-                """print(v7 for v8 in v9)""",
-                # Unary ops...
-                """v = -1 if a < b else -2""",
-                # Returns...
-                """return -1""",
-            )
-        fails = 0
-        for i, contents in enumerate(table):
-            description = f"{tag} part {i}"
-            contents, tokens, tree = self.make_data(contents, description=description)
-            expected = self.blacken(contents)
-            results = self.beautify(contents, tokens, tree, filename=description)
-            if results != expected:  # pragma: no cover
-                fails += 1
-                print('')
-                print(
-                    f"TestOrange.test_one_line_pet_peeves: FAIL {fails}\n"
-                    f"  contents: {contents.rstrip()}\n"
-                    f"     black: {expected.rstrip()}\n"
-                    f"    orange: {results.rstrip() if results else 'None'}"
-                )
-        self.assertEqual(fails, 0)
-
-    # @+node:ekr.20220327135448.1: *4* TestOrange.test_relative_imports
-    def test_relative_imports(self):
-        # #2533.
-        contents = self.prep(
-            """
-            from .module1 import w
-            from . module2 import x
-            from ..module1 import y
-            from .. module2 import z
-            from . import a
-            from.import b
-            from leo.core import leoExternalFiles
-            import leo.core.leoGlobals as g
-        """
-        )
-
-        expected = self.prep(
-            """
-            from .module1 import w
-            from .module2 import x
-            from ..module1 import y
-            from ..module2 import z
-            from . import a
-            from . import b
-            from leo.core import leoExternalFiles
-            import leo.core.leoGlobals as g
-        """
-        )
-
-        contents, tokens, tree = self.make_data(contents)
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(expected, results)
-
-    # @+node:ekr.20200210050646.1: *4* TestOrange.test_return
-    def test_return(self):
-        contents = """return []"""
-        expected = self.blacken(contents)
-        contents, tokens, tree = self.make_data(contents)
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200107174742.1: *4* TestOrange.test_single_quoted_string
-    def test_single_quoted_string(self):
-        contents = """print('hi')"""
-        # blacken suppresses string normalization.
-        expected = self.blacken(contents)
-        contents, tokens, tree = self.make_data(contents)
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200117180956.1: *4* TestOrange.test_split_lines
-    def test_split_lines(self):
-        line_length = 40  # For testing.
-        table = (
-            # 1234567890x1234567890x1234567890x1234567890x
-            """
-    if 1:
-        print('1111111111', '2222222222', '3333333333')
-    """,
-            """print('aaaaaaaaaaaaa', 'bbbbbbbbbbbbbb', 'cccccc')""",
-            """print('aaaaaaaaaaaaa', 'bbbbbbbbbbbbbb', 'cccccc', 'ddddddddddddddddd')""",
-        )
-        fails = 0
-        for contents in table:
-            contents, tokens, tree = self.make_data(contents)
-            if 0:
-                dump_tokens(tokens)
-                # dump_tree(tokens, tree)
-            expected = self.blacken(contents, line_length=line_length)
-            results = self.beautify(
-                contents,
-                tokens,
-                tree,
-                max_join_line_length=line_length,
-                max_split_line_length=line_length,
-            )
-            message = f"\n  contents: {contents!s}\n     black: {expected!s}\n    orange: {results!s}"
-            if results != expected:  # pragma: no cover
-                fails += 1
-                print(f"Fail: {fails}\n{message}")
-        self.assertEqual(fails, 0)
-
-    # @+node:ekr.20200210073227.1: *4* TestOrange.test_split_lines_2
-    def test_split_lines_2(self):
-        line_length = 40  # For testing.
-
-        # Different from how black handles things.
-        contents = """
-    if not any([z.kind == 'lt' for z in line_tokens]):
-        return False
-    """
-
-        expected = self.prep(
-            """
-            if not any(
-                [z.kind == 'lt' for z in line_tokens]):
-                return False
-        """
-        )
-
-        fails = 0
-        contents, tokens, tree = self.make_data(contents)
-        results = self.beautify(
-            contents,
-            tokens,
-            tree,
-            max_join_line_length=line_length,
-            max_split_line_length=line_length,
-        )
-        message = f"\n  contents: {contents!r}\n  expected: {expected!r}\n       got: {results!r}"
-        if results != expected:  # pragma: no cover
-            fails += 1
-            print(f"Fail: {fails}\n{message}")
-        self.assertEqual(fails, 0)
-
-    # @+node:ekr.20200219144837.1: *4* TestOrange.test_split_lines_3
-    def test_split_lines_3(self):
-        line_length = 40  # For testing.
-        # Different from how black handles things.
-        contents = """print('eee', ('fffffff, ggggggg', 'hhhhhhhh', 'iiiiiii'), 'jjjjjjj', 'kkkkkk')"""
-
-        # This is a bit different from black, but it's good enough for now.
-        expected = self.prep(
-            """
-            print(
-                'eee',
-                ('fffffff, ggggggg', 'hhhhhhhh', 'iiiiiii'),
-                'jjjjjjj',
-                'kkkkkk',
-            )
-        """
-        )
-        fails = 0
-        contents, tokens, tree = self.make_data(contents)
-        results = self.beautify(
-            contents,
-            tokens,
-            tree,
-            max_join_line_length=line_length,
-            max_split_line_length=line_length,
-        )
-        message = f"\n  contents: {contents!r}\n  expected: {expected!r}\n       got: {results!r}"
-        if results != expected:  # pragma: no cover
-            fails += 1
-            print(f"Fail: {fails}\n{message}")
-        self.assertEqual(fails, 0)
-
-    # @+node:ekr.20220401191253.1: *4* TestOrange.test_star_star_operator
-    def test_star_star_operator(self):
-        # Was tested in pet peeves, but this is more permissive.
-        contents = """a = b ** c"""
-        contents, tokens, tree = self.make_data(contents)
-        # Don't rely on black for this test.
-        # expected = self.blacken(contents)
-        expected = contents
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200119155207.1: *4* TestOrange.test_sync_tokens
-    def test_sync_tokens(self):
-        contents = """if x == 4: pass"""
-        # At present Orange doesn't split lines...
-        expected = """if x == 4: pass"""
-        contents, tokens, tree = self.make_data(contents)
-        expected = self.adjust_expected(expected)
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200209161226.1: *4* TestOrange.test_ternary
-    def test_ternary(self):
-        contents = """print(2 if name == 'class' else 1)"""
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected)
-
-    # @+node:ekr.20200211093359.1: *4* TestOrange.test_verbatim
-    def test_verbatim(self):
-        line_length = 40  # For testing.
-
-        contents = self.prep("""
-    # @@nobeautify
-
-    def addOptionsToParser(self, parser, trace_m):
-
-        add = parser.add_option
-
-        def add_bool(option, help, dest=None):
-            add(option, action='store_true', dest=dest, help=help)
-
-        add_bool('--diff',          'use Leo as an external git diff')
-        # add_bool('--dock',          'use a Qt dock')
-        add_bool('--fullscreen',    'start fullscreen')
-        add_bool('--init-docks',    'put docks in default positions')
-        # Multiple bool values.
-        add('-v', '--version', action='store_true',
-            help='print version number and exit')
-
-    # From leoAtFile.py
-    noDirective     =  1 # not an at-directive.
-    allDirective    =  2 # at-all (4.2)
-    docDirective    =  3 # @doc.
-
-    # @@beautify
-    """)
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents
-        results = self.beautify(
-            contents,
-            tokens,
-            tree,
-            max_join_line_length=line_length,
-            max_split_line_length=line_length,
-        )
-        self.assertEqual(results, expected, msg=contents)
-
-    # @+node:ekr.20200211094209.1: *4* TestOrange.test_verbatim_with_pragma
-    def test_verbatim_with_pragma(self):
-        line_length = 40  # For testing.
-        contents = """
-    # pragma: no beautify
-
-    def addOptionsToParser(self, parser, trace_m):
-
-        add = parser.add_option
-
-        def add_bool(option, help, dest=None):
-            add(option, action='store_true', dest=dest, help=help)
-
-        add_bool('--diff',          'use Leo as an external git diff')
-        # add_bool('--dock',          'use a Qt dock')
-        add_bool('--fullscreen',    'start fullscreen')
-        add_other('--window-size',  'initial window size (height x width)', m='SIZE')
-        add_other('--window-spot',  'initial window position (top x left)', m='SPOT')
-        # Multiple bool values.
-        add('-v', '--version', action='store_true',
-            help='print version number and exit')
-
-    # pragma: beautify
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents
-        results = self.beautify(
-            contents,
-            tokens,
-            tree,
-            max_join_line_length=line_length,
-            max_split_line_length=line_length,
-        )
-        self.assertEqual(results, expected, msg=contents)
-
-    # @+node:ekr.20200729083027.1: *4* TestOrange.verbatim2
-    def test_verbatim2(self):
-        contents = """
-    # @@beautify
-    # @+at Starts doc part
-    # More doc part.
-    # The @c ends the doc part.
-    # @@c
-    """
-        contents, tokens, tree = self.make_data(contents)
-        expected = contents
-        results = self.beautify(contents, tokens, tree)
-        self.assertEqual(results, expected, msg=contents)
 
     # @-others
