+++ ruff.toml
@@ -31,6 +31,7 @@ exclude = [
     "leo/plugins/test/failed_import.py",  # Contains unused import of 'xyzzy'.
     "leo/plugins/pygeotag/pygeotag.py",  # Contains strange errors.
     "leo/test/scriptFile.py",  # Generated by *user* scripts.
+    "leo/test/beautify_node.py",  # Generated from user scripts.
     "leo/unittests/py3_test_grammar.py",  # Contains many errors on purpose.
 ]
